Converted ResponseBodyHelper component to ts
