+++ gui2/src/components/stub/ResponseBodyHelper.vue
@@ -20,7 +20,7 @@
       </div>
     </div>
 
-    <div v-if="responseBodyType === responseBodyTypes.base64">
+    <div v-if="responseBodyType === ResponseBodyType.base64">
       <div>
         <div class="hint">
           You can upload a <strong>file</strong> for use in the Base64 response
@@ -86,23 +86,32 @@
   </div>
 </template>
 
-<script>
+<script lang="ts">
 import {
-  elementDescriptions,
-  responseBodyTypes,
-} from "@/constants/stubFormResources";
-import { computed, onMounted, ref, watch } from "vue";
+  computed,
+  defineComponent,
+  onMounted,
+  type PropType,
+  ref,
+  watch,
+} from "vue";
 import { handleHttpError } from "@/utils/error";
 import { fromBase64, toBase64 } from "@/utils/text";
 import { useMetadataStore } from "@/store/metadata";
-import { useStubFormStore } from "@/store/stubForm";
-import { defineComponent } from "vue";
+import { type SetResponseInput, useStubFormStore } from "@/store/stubForm";
+import {
+  getValues,
+  ResponseBodyType,
+} from "@/domain/stubForm/response-body-type";
+import type { MetadataModel } from "@/domain/metadata/metadata-model";
+import type { FileUploadedModel } from "@/domain/file-uploaded-model";
+import { elementDescriptions } from "@/domain/stubForm/element-descriptions";
 
 export default defineComponent({
   name: "ResponseBodyHelper",
   props: {
     presetResponseBodyType: {
-      type: String,
+      type: String as PropType<ResponseBodyType>,
     },
   },
   setup(props) {
@@ -110,29 +119,28 @@ export default defineComponent({
     const stubFormStore = useStubFormStore();
 
     // Refs
-    const codeEditor = ref(null);
+    const codeEditor = ref<any>();
 
     // Data
-    const responseBodyType = ref("");
+    const responseBodyType = ref<ResponseBodyType>(ResponseBodyType.text);
     const responseBody = ref("");
-    const enableDynamicMode = ref(null);
+    const enableDynamicMode = ref<boolean | undefined>();
     const showBase64TextInput = ref(false);
-    const responseBodyTypeItems = Object.keys(responseBodyTypes).map(
-      (k) => responseBodyTypes[k]
-    );
-    const metadata = ref(null);
+    const responseBodyTypeItems = getValues();
+    const metadata = ref<MetadataModel>();
     const selectedVariableHandler = ref("");
     const cmOptions = ref({
       tabSize: 4,
-      mode: "",
+      mode: "" as any,
       lineNumbers: true,
       line: true,
+      htmlMode: false,
     });
-    let setInputTimeout = null;
+    let setInputTimeout: any;
 
     // Computed
     const showDynamicModeRow = computed(
-      () => responseBodyType.value !== responseBodyTypes.base64
+      () => responseBodyType.value !== ResponseBodyType.base64
     );
     const showVariableParsers = computed(
       () => showDynamicModeRow.value && enableDynamicMode.value
@@ -161,38 +169,42 @@ export default defineComponent({
     });
     const showResponseBody = computed(
       () =>
-        responseBodyType.value !== responseBodyTypes.base64 ||
+        responseBodyType.value !== ResponseBodyType.base64 ||
         showBase64TextInput.value
     );
 
     // Methods
-    const onUploaded = (file) => {
+    const onUploaded = (file: FileUploadedModel) => {
       const regex = /^data:(.+);base64,(.*)$/;
       const matches = file.result.match(regex);
       const contentType = matches[1];
       const body = matches[2];
       stubFormStore.setResponseContentType(contentType);
       stubFormStore.setResponseBody({
-        type: responseBodyTypes.base64,
+        type: ResponseBodyType.base64,
         body,
-      });
+      } as SetResponseInput);
       responseBody.value = body;
       stubFormStore.closeFormHelper();
       showBase64TextInput.value = false;
     };
     const insertVariableHandler = () => {
       if (codeEditor.value && codeEditor.value.replaceSelection) {
-        const handler = metadata.value.variableHandlers.find(
-          (h) => h.name === selectedVariableHandler.value
-        );
-        setTimeout(() => (selectedVariableHandler.value = ""), 10);
-        codeEditor.value.replaceSelection(handler.example);
+        if (metadata.value) {
+          const handler = metadata.value.variableHandlers.find(
+            (h) => h.name === selectedVariableHandler.value
+          );
+          if (handler) {
+            setTimeout(() => (selectedVariableHandler.value = ""), 10);
+            codeEditor.value.replaceSelection(handler.example);
+          }
+        }
       }
     };
     const insert = () => {
       let responseBodyResult = responseBody.value;
-      if (responseBodyType.value === responseBodyTypes.base64) {
-        responseBodyResult = toBase64(responseBodyResult);
+      if (responseBodyType.value === ResponseBodyType.base64) {
+        responseBodyResult = toBase64(responseBodyResult) as string;
       }
 
       stubFormStore.setResponseBody({
@@ -211,7 +223,7 @@ export default defineComponent({
       responseBodyType.value =
         props.presetResponseBodyType || stubFormStore.getResponseBodyType;
       let currentResponseBody = stubFormStore.getResponseBody;
-      if (responseBodyType.value === responseBodyTypes.base64) {
+      if (responseBodyType.value === ResponseBodyType.base64) {
         const decodedBase64 = fromBase64(currentResponseBody);
         if (decodedBase64) {
           currentResponseBody = decodedBase64;
@@ -233,15 +245,15 @@ export default defineComponent({
       cmOptions.value.htmlMode = false;
       cmOptions.value.mode = "";
       switch (responseBodyType.value) {
-        case responseBodyTypes.html:
+        case ResponseBodyType.html:
           cmOptions.value.htmlMode = true;
           cmOptions.value.mode = "text/html";
           break;
-        case responseBodyTypes.xml:
+        case ResponseBodyType.xml:
           cmOptions.value.htmlMode = false;
           cmOptions.value.mode = "application/xml";
           break;
-        case responseBodyTypes.json:
+        case ResponseBodyType.json:
           cmOptions.value.mode = { name: "javascript", json: true };
           break;
       }
@@ -259,8 +271,6 @@ export default defineComponent({
       enableDynamicMode,
       responseBodyTypeItems,
       showDynamicModeRow,
-      responseBodyTypes,
-      elementDescriptions,
       showBase64TextInput,
       showVariableParsers,
       selectedVariableHandler,
@@ -274,6 +284,8 @@ export default defineComponent({
       showResponseBodyTypeDropdown,
       cmOptions,
       codeEditor,
+      ResponseBodyType,
+      elementDescriptions,
     };
   },
 });
