--- gui2/src/constants/featureFlagType.js
@@ -1,3 +1,0 @@
-export const FeatureFlagType = {
-  Authentication: "Authentication",
-};
