+++ gui2/src/domain/metadata/feature-flag-type.ts
@@ -1,0 +1,3 @@
+export enum FeatureFlagType {
+  Authentication = "Authentication",
+}
