+++ gui2/src/domain/metadata/feature-result-model.ts
@@ -1,0 +1,6 @@
+import type { FeatureFlagType } from "@/domain/metadata/feature-flag-type";
+
+export interface FeatureResultModel {
+  featureFlag: FeatureFlagType;
+  enabled: boolean;
+}
