+++ gui2/src/domain/metadata/metadata-model.ts
@@ -1,0 +1,6 @@
+import type { VariableHandlerModel } from "@/domain/metadata/variable-handler-model";
+
+export interface MetadataModel {
+  version: string;
+  variableHandlers: VariableHandlerModel[];
+}
