+++ gui2/src/domain/metadata/variable-handler-model.ts
@@ -1,0 +1,5 @@
+export interface VariableHandlerModel {
+  name: string;
+  fullName: string;
+  example: string;
+}
