+++ gui2/src/domain/stubForm/response-body-type.ts
@@ -5,3 +5,13 @@ export enum ResponseBodyType {
   html = "HTML",
   base64 = "Base64",
 }
+
+export function getValues(): string[] {
+  return [
+    ResponseBodyType.text,
+    ResponseBodyType.json,
+    ResponseBodyType.xml,
+    ResponseBodyType.html,
+    ResponseBodyType.base64,
+  ];
+}
